Extract acquirer keyselectors from XmlProcessor
