+++ library/src/net/emandates/merchant/library/ISigningKeyProvider.java
@@ -1,5 +1,6 @@
 package net.emandates.merchant.library;
 
+import javax.xml.crypto.KeySelector;
 import java.io.IOException;
 import java.security.KeyStoreException;
 import java.security.NoSuchAlgorithmException;
@@ -8,4 +9,6 @@ import java.security.cert.CertificateException;
 
 public interface ISigningKeyProvider {
     public SigningKeyPair getSigningKeyPair() throws CertificateException, NoSuchAlgorithmException, KeyStoreException, UnrecoverableEntryException, CommunicatorException, IOException;
+    public KeySelector getAcquirerKeySelector();
+    public KeySelector getAlternativeAcquirerKeySelector();
 }
